Add parser table
